requirements.txt
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.10.0
attrdict==2.0.1
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
decorator==5.2.1
einops==0.8.1
einx==0.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.12.19
frozendict==2.4.6
fsspec==2025.7.0
future==1.0.0
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.4
modelscope==1.22.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.20.1
onnxruntime-gpu==1.23.2
openai-whisper==20240930
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==7.34.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pynini==2.1.6
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
resampy==0.4.3
rich==14.1.0
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchdiffeq==0.2.5
torchvision==0.22.1
tqdm==4.67.1
transformers==4.49.0
triton==3.3.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
WeTextProcessing==1.0.4.1
x-transformers==2.7.2
